New tests class for Polygon + correction of method polygonArea() in GeometryUtils to ensure that the computed area is positive.
